+++ client/src/app/app.module.ts
@@ -15,9 +15,9 @@ import {AuthGuard} from "./auth/auth-guard.service";
 import {AuthHttp, AuthConfig} from "angular2-jwt";
 import {TemperatureViewComponent} from "./temperature-view/temperature-view.component";
 import {DevicesSetupComponent} from "./devices-setup/devices-setup.component";
-import {BlindsSetupComponent} from "./blinds-setup/blinds-setup.component";
-import {HumiditySetupComponent} from "./humidity-setup/humidity-setup.component";
-import {TemperatureSetupComponent} from "./temperature-setup/temperature-setup.component";
+import {BlindsSetupComponent} from "./devices-setup/blinds-setup/blinds-setup.component";
+import {HumiditySetupComponent} from "./devices-setup/humidity-setup/humidity-setup.component";
+import {TemperatureSetupComponent} from "./devices-setup/temperature-setup/temperature-setup.component";
 
 // TODO: in welches File müsste diese Methode?
 export function authHttpServiceFactory(http: Http, options: RequestOptions) {
