+++ client/src/app/devices-setup/blinds-setup/blinds-setup.component.scss
@@ -23,7 +23,7 @@ main {
 }
 
 .deviceFormLabel {
-  flex: 0 0 8em;
+  flex: 0 0 10em;
 }
 
 .deviceFormValue {
