+++ client/src/app/devices-setup/humidity-setup/humidity-setup.component.ts
@@ -1,10 +1,10 @@
 import {Component, OnInit} from '@angular/core';
 import {Router} from "@angular/router";
-import {GenericService} from "../remote/generic.service";
+import {GenericService} from "../../remote/generic.service";
 import {AuthHttp} from "angular2-jwt";
-import {ClientSocketService} from "../remote/client-socket.service";
+import {ClientSocketService} from "../../remote/client-socket.service";
 
-import {HumidityDevice, HumidityDeviceCharacteristics, Port, portName} from '../device-pool';
+import {HumidityDevice, HumidityDeviceCharacteristics, Port, portName} from '../../device-pool';
 
 
 @Component({
@@ -13,7 +13,7 @@ import {HumidityDevice, HumidityDeviceCharacteristics, Port, portName} from '../
   styleUrls: ['humidity-setup.component.scss']
 })
 export class HumiditySetupComponent implements OnInit {
-
+  headerTitle: string = 'FEUCHTIGKEIT-SETUP';
   devices: HumidityDevice[] = [];
   device: HumidityDevice;
   selectedDevice: HumidityDevice;
