+++ client/src/app/devices-setup/temperature-setup/temperature-setup.component.html
@@ -1,7 +1,7 @@
 <header>
   <img id="backIcon" class="headerAction" src="app/images/back.svg" alt="Back Icon" title="Back Icon"
        (click)="backClicked()">
-  <h1 class="headerTitle">FEUCHTIGKEIT-SETUP</h1>
+  <h1 class="headerTitle">{{headerTitle}}</h1>
   <div class="contextHeaderActionDiv">
     <img id="addIcon" class="headerAction contextHeaderAction" src="app/images/add.svg" alt="Add Icon"
          title="Add Icon" (click)="addClicked()">
