+++ client/src/app/devices-setup/temperature-setup/temperature-setup.component.ts
@@ -1,10 +1,10 @@
 import {Component, OnInit} from '@angular/core';
 import {Router} from "@angular/router";
-import {GenericService} from "../remote/generic.service";
+import {GenericService} from "../../remote/generic.service";
 import {AuthHttp} from "angular2-jwt";
-import {ClientSocketService} from "../remote/client-socket.service";
+import {ClientSocketService} from "../../remote/client-socket.service";
 
-import {TemperatureDevice, TemperatureDeviceCharacteristics, Port, portName} from '../device-pool';
+import {TemperatureDevice, TemperatureDeviceCharacteristics, Port, portName} from '../../device-pool';
 
 
 @Component({
@@ -13,7 +13,7 @@ import {TemperatureDevice, TemperatureDeviceCharacteristics, Port, portName} fro
   styleUrls: ['temperature-setup.component.scss']
 })
 export class TemperatureSetupComponent implements OnInit {
-
+  headerTitle: string = 'TEMPERATUR-SETUP';
   devices: TemperatureDevice[] = [];
   device: TemperatureDevice;
   selectedDevice: TemperatureDevice;
